Production/functions/recieve_twilio_sms/requirements.txt
jsonpickle==1.3
aws-xray-sdk==2.4.3
twilio==6.45.3
boto3==1.15.6
python-dotenv==0.14.0
requests==2.24.0
hashids==1.3.1
